Moved HostingPlatform to separate folder
